Update: restructure project layout.

+++ tests/ManagedETWTests/describe_Asserts.cs
@@ -120,7 +120,7 @@ namespace EtwTestsCS
 
             var provider = new Provider(WinINetEvent.ProviderId);
             provider.OnEvent +=
-                e => Assert.AreEqual(data, e.GetInt16(prop));
+                e => Assert.AreEqual(data, (ushort)e.GetInt16(prop));
 
             trace.Enable(provider);
             proxy.PushEvent(WinINetEvent.CreateRecord(
